+++ src/test/java/com/liberty52/product/service/controller/guest/OrderGuestControllerTest.java
@@ -1,11 +1,10 @@
-package com.liberty52.product.service.controller;
+package com.liberty52.product.service.controller.guest;
 
 import com.liberty52.product.global.exception.external.ErrorResponse;
 import com.liberty52.product.global.exception.external.RestExceptionHandler;
 import com.liberty52.product.global.exception.external.badrequest.CannotAccessOrderException;
 import com.liberty52.product.service.applicationservice.OrderCreateService;
 import com.liberty52.product.service.applicationservice.OrderRetrieveService;
-import com.liberty52.product.service.controller.guest.OrderGuestController;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
@@ -19,7 +18,6 @@ import java.time.LocalDate;
 
 import static com.liberty52.product.service.utils.MockConstants.*;
 import static com.liberty52.product.service.utils.MockFactory.createMockOrderDetailRetrieveResponse;
-import static com.liberty52.product.service.utils.MockFactory.createMockOrderRetrieveResponseList;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
@@ -27,64 +25,33 @@ import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@WebMvcTest(value = {OrderRetrieveController.class, OrderGuestController.class, RestExceptionHandler.class})
-class OrderRetrieveControllerTest {
-
-//    @InjectMocks
-//    OrderRetrieveController orderRetrieveController;
-//    @InjectMocks
-//    GuestOrderRetrieveController guestOrderRetrieveController;
-
-    @MockBean
-    OrderRetrieveService orderRetrieveService;
+@WebMvcTest(value = {OrderGuestController.class,RestExceptionHandler.class})
+public class OrderGuestControllerTest {
 
     @MockBean
     OrderCreateService orderCreateService;
 
     @MockBean
-    RestExceptionHandler exceptionHandler;
+    OrderRetrieveService orderRetrieveService;
 
     @Autowired
     MockMvc mockMvc;
 
     final String ORDER_URL = "/orders";
     final String GUEST_PREFIX = "/guest";
+    @MockBean
+    RestExceptionHandler exceptionHandler;
 
     @Test
-    void retrieveOrderForList () throws Exception{
+    void retrieveGuestOrderDetail () throws Exception{
         //given
-        given(orderRetrieveService.retrieveOrders(MOCK_AUTH_ID))
-                .willReturn(createMockOrderRetrieveResponseList());
-        //when
-
-        mockMvc.perform(get(ORDER_URL)
-                .header(HttpHeaders.AUTHORIZATION,MOCK_AUTH_ID ))
-        //then
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.length()").value(MOCK_LIST_SIZE))
-                .andExpect(jsonPath("$.[0].orderId").value(MOCK_ORDER_ID))
-                .andExpect(jsonPath("$.[0].orderDate").value(LocalDate.now().toString()))
-                .andExpect(jsonPath("$.[0].orderStatus").value(MOCK_ORDER_STATUS_ORDERED.name()))
-                .andExpect(jsonPath("$.[0].address").value(MOCK_ADDRESS))
-                .andExpect(jsonPath("$.[0].receiverEmail").value(MOCK_RECEIVER_EMAIL))
-                .andExpect(jsonPath("$.[0].receiverPhoneNumber").value(MOCK_RECEIVER_PHONE_NUMBER))
-                .andExpect(jsonPath("$.[0].receiverName").value(MOCK_RECEIVER_NAME))
-                .andExpect(jsonPath("$.[0].products[0].name").value(MOCK_PRODUCT_NAME))
-                .andExpect(jsonPath("$.[0].products[0].quantity").value(MOCK_QUANTITY))
-                .andExpect(jsonPath("$.[0].products[0].price").value(MOCK_PRICE))
-                .andDo(print());
-    }
-
-    @Test
-    void retrieveOrderDetail () throws Exception{
-        //given
-        given(orderRetrieveService.retrieveOrderDetail(MOCK_AUTH_ID, MOCK_ORDER_ID))
+        given(orderRetrieveService.retrieveGuestOrderDetail(MOCK_AUTH_ID, MOCK_ORDER_ID))
                 .willReturn(createMockOrderDetailRetrieveResponse());
 
         //when
-        mockMvc.perform(get(ORDER_URL+"/"+MOCK_ORDER_ID)
-                .header(HttpHeaders.AUTHORIZATION,MOCK_AUTH_ID))
-        //then
+        mockMvc.perform(get(GUEST_PREFIX+ORDER_URL+"/"+MOCK_ORDER_ID)
+                        .header(HttpHeaders.AUTHORIZATION,MOCK_AUTH_ID))
+                //then
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.orderId").value(MOCK_ORDER_ID))
                 .andExpect(jsonPath("$.orderDate").value(LocalDate.now().toString()))
@@ -104,19 +71,19 @@ class OrderRetrieveControllerTest {
     }
 
     @Test
-    void retrieveOrderDetail_throw_cannot_access () throws Exception{
+    void retrieveGuestOrderDetail_throw_cannot_access () throws Exception{
         //given
-        given(orderRetrieveService.retrieveOrderDetail(MOCK_AUTH_ID, MOCK_ORDER_ID))
+        given(orderRetrieveService.retrieveGuestOrderDetail(MOCK_AUTH_ID, MOCK_ORDER_ID))
                 .willThrow(CannotAccessOrderException.class);
         given(exceptionHandler.handleGlobalException(any(),any()))
                 .willReturn(
                         ResponseEntity.status(HttpStatus.BAD_REQUEST)
-                        .body(ErrorResponse.createErrorResponse(new CannotAccessOrderException(), ORDER_URL+"/"+MOCK_ORDER_ID))
+                                .body(ErrorResponse.createErrorResponse(new CannotAccessOrderException(), ORDER_URL+"/"+MOCK_ORDER_ID))
                 );
 
         //when         //then
-        mockMvc.perform(get(ORDER_URL+"/"+MOCK_ORDER_ID)
-                .header(HttpHeaders.AUTHORIZATION,MOCK_AUTH_ID))
+        mockMvc.perform(get(GUEST_PREFIX+ORDER_URL+"/"+MOCK_ORDER_ID)
+                        .header(HttpHeaders.AUTHORIZATION,MOCK_AUTH_ID))
                 .andExpect(status().isBadRequest())
                 .andDo(print());
     }
